[15.0][MIG] stock_request_kanban
