+++ stock_request_kanban/static/src/xml/stock_request_kanban_scan.xml
@@ -1,6 +1,10 @@
 <?xml version="1.0" encoding="utf-8" ?>
 <templates id="template" xml:space="preserve">
-    <button t-name="StockRequestKanban.Buttons" class="btn-primary btn" type="button">
+    <button
+        t-name="StockRequestKanban.Buttons"
+        class="btn-primary btn mr-1"
+        type="button"
+    >
         Scan Kanban
     </button>
 </templates>
