+++ stock_request_kanban/views/stock_request_kanban_views.xml
@@ -7,7 +7,6 @@
         <field name="model">stock.request.kanban</field>
         <field name="arch" type="xml">
             <tree
-                string="Stock Requests"
                 decoration-muted="active == 'false'"
                 decoration-bf="message_needaction==True"
             >
